requirements.txt
psycopg2 == 2.9.5
django-cron == 0.6.0
Pillow == 9.5.0
mysqlclient == 2.1.1